resolving refactor issue
