+++ resqpy/organize/__init__.py
@@ -1,3 +1,7 @@
+"""
+resqpy.organize: Package for RESQML organizational objects.
+"""
+
 __all__ = [
     'OrganizationFeature',
     'GeobodyFeature',
@@ -13,9 +17,9 @@ __all__ = [
     'GeneticBoundaryFeature',
 ]
 
-from resqpy.organize.organize_functions import (equivalent_extra_metadata, alias_for_attribute,
-                                                extract_has_occurred_during, equivalent_chrono_pairs,
-                                                create_xml_has_occurred_during)
+from resqpy.organize._utils import (equivalent_extra_metadata, alias_for_attribute,
+                                    extract_has_occurred_during, equivalent_chrono_pairs,
+                                    create_xml_has_occurred_during)
 from .boundary_feature import BoundaryFeature
 from .earth_model_interpretation import EarthModelInterpretation
 from .fault_interpretation import FaultInterpretation
