+++ resqpy/organize/fault_interpretation.py
@@ -1,7 +1,9 @@
+"""Class for RESQML Fault Interpretation organizational objects."""
+
 import math as maths
 
-from .organize_functions import (equivalent_extra_metadata, alias_for_attribute, extract_has_occurred_during,
-                                 equivalent_chrono_pairs, create_xml_has_occurred_during)
+from ._utils import (equivalent_extra_metadata, alias_for_attribute, extract_has_occurred_during,
+                     equivalent_chrono_pairs, create_xml_has_occurred_during)
 
 import resqpy.olio.uuid as bu
 import resqpy.olio.xml_et as rqet
